requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
